Rename TestChart to TempChart: temprature
